+++ trabalho-pratico/include/data/schemas/reservation.h
@@ -33,12 +33,12 @@ typedef struct reservation Reservation;
  * @param comment Comment.
  * @return New reservation.
  */
-Reservation* create_reservation(unsigned short id, char* user_id,
-                                char* hotel_id, char* hotel_name,
-                                int hotel_stars, int city_tax, char* address,
-                                Timestamp begin_date, Timestamp end_date,
-                                int price_per_night, bool includes_breakfast,
-                                char* room_details, int rating, char* comment);
+Reservation* create_reservation(unsigned int id, char* user_id, char* hotel_id,
+                                char* hotel_name, int hotel_stars, int city_tax,
+                                char* address, Timestamp begin_date,
+                                Timestamp end_date, int price_per_night,
+                                bool includes_breakfast, char* room_details,
+                                int rating, char* comment);
 
 /**
  * @brief Frees reservation.
@@ -68,14 +68,14 @@ bool reservation_invalid_association(void* database, char* user_id);
  * @param reservation Reservation.
  * @return Reservation's id.
  */
-unsigned short reservation_get_id(Reservation* reservation);
+unsigned int reservation_get_id(Reservation* reservation);
 
 /**
  * @brief Sets reservation's id.
  * @param reservation Reservation.
  * @param id Id.
  */
-void reservation_set_id(Reservation* reservation, unsigned short id);
+void reservation_set_id(Reservation* reservation, unsigned int id);
 
 /**
  * @brief Gets reservation's user id.
